+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableApi.java
@@ -828,7 +828,6 @@ public class IterableApi {
             sharedInstance.inAppManager = new IterableInAppManager(
                     sharedInstance,
                     sharedInstance.config.inAppHandler,
-                    sharedInstance.config.inAppDisplayHandler,
                     sharedInstance.config.inAppDisplayInterval,
                     sharedInstance.config.useInMemoryStorageForInApps);
         }